+++ src/Model/Config/Communication.php
@@ -9,7 +9,6 @@ use Omikron\FactFinder\Oxid\Export\Filter\TextFilter;
 use OxidEsales\Eshop\Application\Controller\FrontendController;
 use OxidEsales\Eshop\Application\Model\Category;
 use OxidEsales\Eshop\Core\Registry;
-use RuntimeException;
 
 class Communication implements ParametersSourceInterface
 {
@@ -33,6 +32,7 @@ class Communication implements ParametersSourceInterface
         $category = $this->view->getActiveCategory();
         $session  = Registry::getSession();
 
+        var_dump($this->view->getActiveLangAbbr());
         $params = [
             'url'                         => $this->getConfig('ffServerUrl'),
             'version'                     => $this->getConfig('ffApiVersion'),
@@ -118,7 +118,7 @@ class Communication implements ParametersSourceInterface
         $channels = $this->getConfig('ffChannel');
 
         if (!isset($channels[$langAbbr])) {
-            throw new RuntimeException("No channel for used language: $langAbbr");
+            $channels[$langAbbr] = '';
         }
 
         return $channels[$langAbbr];